backend/guardian/requirements.txt

Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-SocketIO==5.3.6
Werkzeug==2.3.7
gunicorn==21.2.0
eventlet==0.33.3
python-socketio==5.8.0
python-engineio==4.7.1
requests==2.31.0
GitPython==3.1.37
psycopg2-binary==2.9.8
python-dotenv==1.0.0
